view cleanup
