+++ Stitch/Graph/Sidebar/View/SidebarListView.swift
@@ -22,52 +22,9 @@ let SIDEBAR_LIST_ITEM_ROW_COLORED_AREA_HEIGHT: CGFloat = 32.0
 let SIDEBAR_LIST_ITEM_FONT: Font = stitchFont(18)
 #endif
 
-// TODO: move
-enum ProjectSidebarTab: String, Identifiable, CaseIterable {
-    case layers = "Layers"
-    case assets = "Assets"
-}
-
-extension ProjectSidebarTab {
-    var id: String {
-        self.rawValue
-    }
-    
-    var iconName: String {
-        switch self {
-        case .layers:
-            return "square.3.layers.3d.down.left"
-        case .assets:
-            return "folder"
-        }
-    }
-    
-//    @ViewBuilder @MainActor
-//    func content(graph: GraphState,
-//                 syncStatus: iCloudSyncStatus) -> some View {
-//        switch self {
-//        case .layers:
-//            LayersSidebarView(graph: graph,
-//                              syncStatus: syncStatus)
-//        case .assets:
-//            Text("Assets")
-//        }
-//    }
-//    
-    var viewModelType: any ProjectSidebarObservable.Type {
-        switch self {
-        case .layers:
-            return LayersSidebarViewModel.self
-        default:
-            fatalError()
-        }
-    }
-}
-
 struct SidebarListView: View {
     static let tabs = ["Layers", "Assets"]
     @State private var currentTab = ProjectSidebarTab.layers.rawValue
-//    @Bindable var layersViewModel: LayersSidebarViewModel
     
     @Bindable var graph: GraphState
     let syncStatus: iCloudSyncStatus
@@ -101,10 +58,6 @@ struct SidebarListView: View {
                 }
             }
         }
-        // TODO: see note in `DeriveSidebarList`
-//        .onChange(of: graph.nodes.keys.count) {
-//            self.graph.layersSidebarViewModel.
-//        }
     }
 }
 
@@ -128,14 +81,6 @@ struct SidebarListScrollView<SidebarObservable>: View where SidebarObservable: P
                               syncStatus: syncStatus)
         }
     }
-    
-//    var allItems: [SidebarObservable.ItemViewModel] {
-//        if let draggedItem = sidebarViewModel.currentItemDragged {
-//            return [draggedItem] + sidebarViewModel.items
-//        }
-//        
-//        return sidebarViewModel.items
-//    }
     
     // Note: sidebar-list-items is a flat list;
     // indentation is handled by calculated indentations.
@@ -187,15 +132,10 @@ struct SidebarListScrollView<SidebarObservable>: View where SidebarObservable: P
         .animation(.spring(), value: selections)
 #endif
         // TODO: remove some of these animations ?
-        .animation(.spring(), value: isBeingEdited)
-//        .animation(.spring(), value: sidebarViewModel.proposedGroup)
-//        .animation(.spring(), value: sidebarDeps)
-//        .animation(.easeIn, value: sidebarViewModel.items)
-        
+        .animation(.spring(), value: isBeingEdited)        
         .onChange(of: isBeingEdited) { _, newValue in
             // This handler enables all animations
             isBeingEditedAnimated = newValue
-//            self.sidebarViewModel.editModeToggled(to: isBeingEdited)
         }
     }
 
@@ -221,285 +161,3 @@ struct SidebarListScrollView<SidebarObservable>: View where SidebarObservable: P
 //        .opacity(0)
 //    }
 }
-
-// TODO: move
-import StitchViewKit
-import OrderedCollections
-
-protocol ProjectSidebarObservable: AnyObject, Observable where ItemViewModel.ID == EncodedItemData.ID,
-                                                               Self.ItemViewModel.SidebarViewModel == Self {
-//                                                               ExcludedGroups: Equatable {
-    associatedtype ItemViewModel: SidebarItemSwipable
-    associatedtype EncodedItemData: StitchNestedListElement
-
-//    typealias ItemData = ItemViewModel.Item
-    typealias ItemID = ItemViewModel.ID
-    typealias SidebarSelectionState = SidebarSelectionObserver<ItemID>
-    typealias SidebarGroupsDict = OrderedDictionary<Self.ItemID, [Self.ItemID]>
-    typealias ExcludedGroups = [ItemID: [ItemViewModel]]
-//    typealias HorizontalDrag = SidebarCursorHorizontalDrag<ItemViewModel>
-    
-    var isEditing: Bool { get set }
-    var items: [ItemViewModel] { get set }
-    
-//    var proposedGroup: ItemID? { get set }
-//    var cursorDrag: Double? { get set }
-    
-    var selectionState: SidebarSelectionState { get set }
-
-    var activeSwipeId: ItemID? { get set }
-    var activeGesture: SidebarListActiveGesture<ItemID> { get set }
-    var implicitlyDragged: Set<ItemID> { get set }
-    var currentItemDragged: Self.ItemID? { get set }
-    
-    var graphDelegate: GraphState? { get set }
-    
-//    init(from encodedData: [Self.EncodedItemData])
-//    func editModeToggled(to isEditing: Bool)
-    func canBeGrouped() -> Bool
-    func canUngroup() -> Bool
-//    func canDuplicate() -> Bool
-    
-//    @MainActor
-//    func didGroupExpand(_ id: ItemID)
-//    @MainActor func sidebarListItemGroupOpened(openedParent: ItemID)
-
-    func sidebarGroupCreated()
-    
-    @MainActor
-    func sidebarGroupUncreatedViaEditMode(groupId: Self.ItemID, children: [Self.ItemID])
-    
-    func didItemsDelete(ids: Set<ItemID>)
-}
-
-extension ProjectSidebarObservable {
-    var proposedGroup: Self.ItemViewModel? {
-        guard let currentItemDragged = self.currentItemDragged else { return nil }
-        return self.items.get(currentItemDragged)?.parentDelegate
-    }
-    
-//    init(from encodedData: [Self.EncodedItemData]) {
-//        self.sync(from: encodedData)
-//    }
-    
-    // the [parentId: child-ids] that are not currently shown
-//    @MainActor var excludedGroups: ExcludedGroups {
-//        let itemsDict = self.items.reduce(into: [Self.ItemID : Self.ItemViewModel]()) { result, item in
-//            result.updateValue(item, forKey: item.id)
-//        }
-//        
-//        let orderedEncodedData = self.createdOrderedEncodedData()
-//        
-//        return self.items.reduce(into: ExcludedGroups()) { result, item in
-//            let isExpandedInSidebar = item.isExpandedInSidebar ?? true
-//            
-//            if item.isGroup && isExpandedInSidebar {
-//                guard let encodedData = orderedEncodedData.getSidebarLayerData(item.id),
-//                      let children = encodedData.children else {
-//                    fatalErrorIfDebug()
-//                    return
-//                }
-//                
-//                let childrenViewModels: [Self.ItemViewModel] = children.compactMap { child in
-//                    guard let viewModel = itemsDict.get(child.id) else {
-//                        fatalErrorIfDebug()
-//                        return nil
-//                    }
-//                    return viewModel
-//                }
-//                
-//                result.updateValue(childrenViewModels, forKey: item.id)
-//            }
-//        }
-//    }
-    
-//    var expandedSidebarItems: Set<ItemID> {
-//        
-//    }
-    
-    // groups currently opened or closed;
-    // an item's id is added when its group closed,
-    // removed when its group opened;
-    // NOTE: a supergroup parent closing/opening does NOT affect a subgroup's closed/open status
-//    var collapsedGroups: Set<ItemID> {
-//        self.items.compactMap {
-//            if $0.isExpandedInSidebar ?? false {
-//                return $0.id
-//            }
-//            return nil
-//        }
-//        .toSet
-//    }
-    
-    var inspectorFocusedLayers: InspectorFocusedData<ItemID> {
-        get {
-            self.selectionState.inspectorFocusedLayers
-        }
-        set(newValue) {
-            self.selectionState.inspectorFocusedLayers = newValue
-        }
-    }
-    
-    func initializeDelegate(graph: GraphState) {
-        self.graphDelegate = graph
-        
-        self.items.recursiveForEach {
-            $0.sidebarDelegate = self
-        }
-//        self.update(from: orderedEncodedData)
-    }
-    
-    @MainActor func persistSidebarChanges(encodedData: [Self.EncodedItemData]? = nil) {
-        // Create new encodable data
-        let encodedData: [Self.EncodedItemData] = encodedData ?? self.createdOrderedEncodedData()
-        
-        // Refreshes view
-        self.update(from: encodedData)
-        
-        self.graphDelegate?.encodeProjectInBackground()
-    }
-    
-    @MainActor func createdOrderedEncodedData() -> [Self.EncodedItemData] {
-        self.items.map { item in
-            item.createSchema()
-        }
-    }
-    
-//    @MainActor private func createEncodableItem(for item: Self.ItemViewModel,
-//                                                itemsAtHierarchy: [Self.ItemViewModel]) -> Self.EncodedItemData {
-//        // Child case
-//        guard item.isGroup else {
-//            return .init(id: item.id,
-//                         children: nil,
-//                         isExpandedInSidebar: nil)
-//        }
-//    
-//        // Find children view models and remove from list as to not duplicate
-//        let childrenViewModels = itemsQueue.filter { $0.parentId == item.id }
-//        let childrenIds = childrenViewModels.map(\.id)
-//        itemsQueue.removeAll(where: { childrenIds.contains($0.id) })
-//        
-//        let encodableChildren = childrenViewModels.map { Self.createEncodableItem(for: $0,
-//                                                                                  itemsQueue: &itemsQueue) }
-//        return .init(id: item.id,
-//                     children: encodableChildren,
-//                     isExpandedInSidebar: item.isExpandedInSidebar)
-//    }
-//
-    func update(from encodedData: [Self.EncodedItemData]) {
-        self.sync(from: encodedData)
-    }
-    
-    func sync(from encodedData: [Self.EncodedItemData]) {
-        let existingViewModels = self.items.reduce(into: [Self.ItemID : Self.ItemViewModel]()) { result, viewModel in
-            result.updateValue(viewModel, forKey: viewModel.id)
-        }
-        
-        self.items = self.recursiveSync(elements: encodedData,
-                                        existingViewModels: existingViewModels)
-        self.items.updateSidebarIndices()
-    }
-    
-    func recursiveSync(elements: [Self.EncodedItemData],
-                       existingViewModels: [Self.ItemID : Self.ItemViewModel],
-                       parent: Self.ItemViewModel? = nil) -> [Self.ItemViewModel] {
-        elements.map { element in
-            let viewModel = existingViewModels[element.id] ?? .init(data: element,
-                                                                    parentDelegate: parent,
-                                                                    sidebarViewModel: self)
-            
-            viewModel.update(from: element)
-            
-            guard let children = element.children else {
-                viewModel.children = nil
-                viewModel.isExpandedInSidebar = nil
-                return viewModel
-            }
-            
-            let childrenViewModels = self.recursiveSync(elements: children,
-                                                        existingViewModels: existingViewModels,
-                                                        parent: viewModel)
-            viewModel.children = childrenViewModels
-            return viewModel
-        }
-    }
-    
-//    static func setLocation(rowIndex: Int,
-//                            nestingLevel: Int) -> CGPoint {
-//        .init(x: CUSTOM_LIST_ITEM_INDENTATION_LEVEL * nestingLevel,
-//              y: CUSTOM_LIST_ITEM_VIEW_HEIGHT * rowIndex)
-//    }
-}
-
-@Observable
-final class LayersSidebarViewModel: ProjectSidebarObservable {
-    typealias EncodedItemData = SidebarLayerData
-    
-    var isEditing = false
-    var items: [SidebarItemGestureViewModel] = []
-    var selectionState = SidebarSelectionObserver<NodeId>()
-//    var orderedEncodedData: OrderedSidebarLayers
-    
-    var activeSwipeId: NodeId?
-    var activeGesture: SidebarListActiveGesture<NodeId> = .none
-    var implicitlyDragged = NodeIdSet()
-    var currentItemDragged: NodeId?
-//    var excludedGroups: [NodeId : [SidebarItemGestureViewModel]] = .init()
-//    var expandedSidebarItems: Set<NodeId> = .init()
-//    var proposedGroup: NodeId?
-//    var cursorDrag: SidebarCursorHorizontalDrag<SidebarItemGestureViewModel>?
-//    var collapsedGroups: Set<NodeId> = .init()
-    
-    weak var graphDelegate: GraphState?
-}
-
-//extension LayersSidebarViewModel {
-////    var expandedSidebarItems: Set<SidebarListItemId> {
-////        self.getSidebarExpandedItems()
-////    }
-//    
-//    @MainActor
-//    func didGroupExpand(_ id: NodeId) {
-//        self.sidebarListItemGroupOpened(parentItem: item)
-//    }
-//}
-
-//struct LayersSidebarView: View {
-//    @Bindable var graph: GraphState
-//
-//    let syncStatus: iCloudSyncStatus
-//    
-//    var body: some View {
-//        VStack {
-//            listView
-//            Spacer()
-//            // Note: previously was in an `.overlay(footer, alignment: .bottom)` which now seems unnecessary
-//            SidebarFooterView(sidebarViewModel: graph.layersSidebarViewModel,
-//                              syncStatus: syncStatus)
-//        }
-//    }
-//
-//    
-//    
-//    // HACK for proper width even when sidebar is empty
-//    // TODO: revisit and re-organize UI to avoid this hack
-//    @ViewBuilder @MainActor
-//    var fakeSidebarListItem: some View {
-//
-//        let item = SidebarListItem.fakeSidebarListItem
-//
-//        SidebarListItemSwipeView(
-//            graph: $graph,
-//            item: item,
-//            name: item.layer.value,
-//            layer: .rectangle,
-//            current: .none,
-//            proposedGroup: .none,
-//            isClosed: true,
-//            selection: .none,
-//            isBeingEdited: false,
-//            activeGesture: $activeGesture,
-//            activeSwipeId: $activeSwipeId)
-//            .opacity(0)
-//    }
-//}
